fix: update tags in feature guide and team creation form
- Changed example tags from "家电维修,上门服务" to "家电维修,宠物服务" in TAGS_FEATURE_GUIDE.md and team creation form.
- Updated the logic in TeaTeamAccountGet to include PendingFromTeamAndUserCount for better tracking of pending transfers.
- Modified the HTML templates to reflect changes in pending transfers from both teams and users.
- Adjusted the SQL schema to include balance_after_receipt for better account management.
- Enhanced the invitation word length validation in route_team_member.go.
- Created a new template for handling pending transfers from users.

+++ Route/route_share_func_tools.go
@@ -46,7 +46,7 @@ func generateHTML(w http.ResponseWriter, template_data any, filenames ...string)
 		"mul": func(a, b int) int {
 			return a * b
 		},
-		"sub": func(a, b float64) float64 {
+		"sub": func(a, b int64) int64 {
 			return a - b
 		},
 		"subtract": func(a, b int) int {
